backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "kidango",email = "kidango0627@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-extensions (>=4.1,<5.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
